requirements.txt
bcrypt==4.0.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
decorator==5.1.1
Deprecated==1.2.14
fabric2==3.2.2
Flask==3.0.0
invoke==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy==1.6.1
mypy-extensions==1.0.0
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
typing_extensions==4.8.0
Werkzeug==3.0.1
wrapt==1.15.0
